chore: Next.js 16 업그레이드 및 Turbopack 전환
- next@15.5.9 → next@16.1.1
- Turbopack 기본 사용으로 전환
- turbopack.root 설정 추가
- webpack 설정 제거

서버/클라이언트 코드 분리 (Turbopack 호환):
- groupPostsByCategory → src/lib/utils/blog.ts
- generateHeadingId, extractCodeText → src/lib/utils/mdx-utils.ts
