+++ next.config.js
@@ -5,16 +5,9 @@ const withBundleAnalyzer = require('@next/bundle-analyzer')({
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  webpack: (config, { isServer }) => {
-    // Three.js 관련 의존성 문제 해결
-    if (!isServer) {
-      config.resolve.fallback = {
-        ...config.resolve.fallback,
-        fs: false,
-        path: false,
-      }
-    }
-    return config
+  // Turbopack 설정 (Next.js 16 기본값)
+  turbopack: {
+    root: __dirname,
   },
   images: {
     remotePatterns: [
