+++ package.json
@@ -27,12 +27,12 @@
     "glob": "^11.0.3",
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.544.0",
-    "next": "15",
+    "next": "16",
     "next-mdx-remote": "^5.0.0",
     "next-themes": "^0.4.6",
     "postcss": "^8.5.6",
-    "react": "19",
-    "react-dom": "19",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "reading-time": "^1.5.0",
     "rehype-pretty-code": "^0.14.1",
     "rehype-slug": "^6.0.0",
